[ORDER] fix: 카프카 프로듀서 메시지 변경 & 카프카 파티션 키 적용

+++ micro-commerce-order/src/main/java/com/microcommerce/order/domain/dto/feign/res/ProductResDto.java
@@ -2,7 +2,7 @@ package com.microcommerce.order.domain.dto.feign.res;
 
 public record ProductResDto(Long id,
                             String name,
-                            long price,
+                            int price,
                             String representativeImageUrl,
                             Long sellerId,
                             String sellerName,
